requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-ollama
faiss-cpu
numpy
PyPDF2
python-docx
